Add files via upload
the script automates the task of renaming files in a specified directory. You need to provide the directory path and define the new file name format. The script uses the os module to list the files in the directory and iterate over them. For each file, it generates a new file name based on the provided format and renames the file accordingly using the os.rename() function. The renaming status is printed for each file.